requirements-dev.txt
# Development requirements (lighter, sin PostgreSQL)
# Usa estos para desarrollo local sin Docker

# Core Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic>=2.6.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.9

# Database (Redis solo)
redis==5.0.1
# psycopg2-binary - NO incluido, usar Docker para tests con DB

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.19.0
structlog==24.1.0

# RAG & Embeddings (Sprint 1)
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
faiss-cpu==1.9.0.post1
pypdf==4.0.1
python-docx==1.1.0
tiktoken>=0.7
